example分割
