+++ CMakeLists.txt
@@ -386,7 +386,12 @@ set(CLI_SRC
     ${CPP_DIR}/c_wcf/view.cc
 )
 set(EXAMPLE_SRC
-    ${CPP_DIR}/example/main.cc
+    ${CPP_DIR}/example/value.cc
+    ${CPP_DIR}/example/func.cc
+    ${CPP_DIR}/example/log.cc
+    ${CPP_DIR}/example/view.cc
+    ${CPP_DIR}/example/canvas2d.cc
+    ${CPP_DIR}/example/canvas3d.cc
     ${CPP_DIR}/example/recv.cc
     ${CPP_DIR}/example/c_example.c
 )
